refactoring use cases

+++ domain/src/main/java/com/justplay1/shoppist/interactor/category/DeleteCategory.java
@@ -43,12 +43,16 @@ public class DeleteCategory extends UseCase<Boolean> {
         this.repository = repository;
     }
 
-    public void setData(Collection<CategoryModel> data) {
+    public DeleteCategory init(Collection<CategoryModel> data) {
         this.data = data;
+        return this;
     }
 
     @Override
     protected Observable<Boolean> buildUseCaseObservable() {
+        if (this.data == null) {
+            throw new IllegalArgumentException("init(data) not called, or called with null argument.");
+        }
         return Observable.fromCallable(() -> {
             repository.delete(data);
             return true;
