+++ domain/src/main/java/com/justplay1/shoppist/interactor/goods/DeleteGoods.java
@@ -42,12 +42,16 @@ public class DeleteGoods extends UseCase<Boolean> {
         this.repository = repository;
     }
 
-    public void setData(Collection<ProductModel> data) {
+    public DeleteGoods init(Collection<ProductModel> data) {
         this.data = data;
+        return this;
     }
 
     @Override
     protected Observable<Boolean> buildUseCaseObservable() {
+        if (this.data == null) {
+            throw new IllegalArgumentException("init(data) not called, or called with null argument.");
+        }
         return Observable.fromCallable(() -> {
             repository.delete(data);
             return true;
