+++ domain/src/main/java/com/justplay1/shoppist/interactor/lists/DeleteLists.java
@@ -42,12 +42,16 @@ public class DeleteLists extends UseCase<Boolean> {
         this.repository = repository;
     }
 
-    public void setData(Collection<ListModel> data) {
+    public DeleteLists init(Collection<ListModel> data) {
         this.data = data;
+        return this;
     }
 
     @Override
     protected Observable<Boolean> buildUseCaseObservable() {
+        if (this.data == null) {
+            throw new IllegalArgumentException("init(data) not called, or called with null argument.");
+        }
         return Observable.fromCallable(() -> {
             repository.delete(data);
             for (ListModel list : data) {
