+++ domain/src/main/java/com/justplay1/shoppist/interactor/units/DeleteUnits.java
@@ -43,12 +43,16 @@ public class DeleteUnits extends UseCase<Boolean> {
         this.repository = repository;
     }
 
-    public void setData(Collection<UnitModel> data) {
+    public DeleteUnits init(Collection<UnitModel> data) {
         this.data = data;
+        return this;
     }
 
     @Override
     protected Observable<Boolean> buildUseCaseObservable() {
+        if (this.data == null) {
+            throw new IllegalArgumentException("init(data) not called, or called with null argument.");
+        }
         return Observable.fromCallable(() -> {
             repository.delete(data);
             return true;
